+++ src/HitRecord.hpp
@@ -18,9 +18,9 @@ public:
     Point3D p;
     Vector3D normal;
     std::shared_ptr<IMaterial> material;
-    double t;
-    double u;
-    double v;
+    float t;
+    float u;
+    float v;
     bool front_face;
 
     void set_face_normal(const Ray &ray, const Vector3D &out_normal)
